adjust to not need junit
I only included Junit since I thought Maven needed it to have any
runner for the tests at all.  Turns out that's not true - SureFire
will run tests if they are in the properly named class, with
a proper test method name.  test, as a method, is not sufficient -
so I named it testEverything, which did work.